Fix banana colouring for skins
